DB 함수 분리
